--- a/config.go
+++ b/config.go
@@ -281,5 +281,4 @@ type metricsOptions struct {
 	disableMetrics  bool
 	httpClient      *http.Client
 	headers         http.Header
-	started         *time.Time
 }