--- a/config.go
+++ b/config.go
@@ -21,7 +21,7 @@ type configOption struct {
 	disableMetrics   bool
 	backupPath       string
 	strategies       []strategy.Strategy
-	listener         interface{}
+	listener         any
 	storage          Storage
 	httpClient       *http.Client
 	customHeaders    http.Header
@@ -35,7 +35,7 @@ type ConfigOption func(*configOption)
 // the listener interfaces. If no listener is registered then the user is responsible
 // for draining the various channels on the client. Failure to do so will stop the client
 // from working as the worker routines will be blocked.
-func WithListener(listener interface{}) ConfigOption {
+func WithListener(listener any) ConfigOption {
 	return func(o *configOption) {
 		o.listener = listener
 	}
@@ -266,7 +266,7 @@ type repositoryOptions struct {
 	backupPath      string
 	refreshInterval time.Duration
 	storage         Storage
-	httpClient     	*http.Client
+	httpClient      *http.Client
 	headers         http.Header
 	isStreaming     bool
 }