--- a/internal/constraints/operator_in.go
+++ b/internal/constraints/operator_in.go
@@ -3,6 +3,7 @@ package constraints
 import (
 	"github.com/Unleash/unleash-go-sdk/v5/api"
 	"github.com/Unleash/unleash-go-sdk/v5/context"
+	"slices"
 )
 
 func operatorNotIn(ctx *context.Context, constraint api.Constraint) bool {
@@ -12,11 +13,5 @@ func operatorNotIn(ctx *context.Context, constraint api.Constraint) bool {
 func operatorIn(ctx *context.Context, constraint api.Constraint) bool {
 	contextValue := ctx.Field(constraint.ContextName)
 
-	for _, constraint := range constraint.Values {
-		if contextValue == constraint {
-			return true
-		}
-	}
-
-	return false
+	return slices.Contains(constraint.Values, contextValue)
 }