--- a/internal/strategies/helpers.go
+++ b/internal/strategies/helpers.go
@@ -1,7 +1,7 @@
 package strategies
 
 import (
-	"math/rand"
+	"math/rand/v2"
 	"os"
 	"strconv"
 	"sync"
@@ -70,9 +70,8 @@ type rng struct {
 
 func (r *rng) int() int {
 	r.Lock()
-	n := r.random.Intn(100) + 1
-	r.Unlock()
-	return n
+	defer r.Unlock()
+	return r.random.IntN(100) + 1
 }
 
 func (r *rng) float() float64 {
@@ -81,14 +80,12 @@ func (r *rng) float() float64 {
 
 func (r *rng) string() string {
 	r.Lock()
-	n := r.random.Intn(10000) + 1
-	r.Unlock()
-	return strconv.Itoa(n)
+	defer r.Unlock()
+	return strconv.Itoa(r.random.IntN(10000) + 1)
 }
 
 // newRng creates a new random number generator and uses a mutex
 // internally to ensure safe concurrent reads.
 func newRng() *rng {
-	seed := time.Now().UnixNano() + int64(os.Getpid())
-	return &rng{random: rand.New(rand.NewSource(seed))}
+	return &rng{random: rand.New(rand.NewPCG(uint64(time.Now().UnixNano()), uint64(os.Getpid())))}
 }