## About the changes

This PR does a couple of small things:
- first we make the benchmarks a bit more realistic by actually adding features (makes IsEnabled less of a noop)
- drop usage of reflect by making `every` generic
- initialize allConstraints to its proper size to avoid multiple slice resizings

The changes above bring a measurable performance difference:
```
goos: linux
goarch: amd64
pkg: github.com/Unleash/unleash-go-sdk/v5
cpu: 12th Gen Intel(R) Core(TM) i7-1250U
                           │  0-old.out  │          3-makeslice.out           │
                           │   sec/op    │   sec/op     vs base               │
FeatureToggleEvaluation-12   1.546µ ± 4%   1.455µ ± 4%  -5.86% (p=0.002 n=10)

                           │  0-old.out   │           3-makeslice.out           │
                           │     B/op     │     B/op      vs base               │
FeatureToggleEvaluation-12   1.383Ki ± 0%   1.328Ki ± 0%  -3.95% (p=0.000 n=10)

                           │ 0-old.out  │          3-makeslice.out           │
                           │ allocs/op  │ allocs/op   vs base                │
FeatureToggleEvaluation-12   13.00 ± 0%   11.00 ± 0%  -15.38% (p=0.000 n=10)
```

Additionally, we also do a couple of minor internal cleanups:
- switch to math/rand/v2
- modernize with golang.org/x/tools/gopls/internal/analysis/modernize/cmd/modernize
- move more code to using stdlib
- do not ignore errors in storage.Reset

### Important files

The biggest impact is caused by the change in `utils.go`, in the `every` function.
Followed by the slice initialization in `client.go`.

## Discussion points
Since there's a lot going on in the PR, one could argue for splitting it into smaller ones.