@costela 2025-09-30
@FredrikOseberg Sure!

- The first commit (26727f4) just gives us a more realistic baseline for benchmarking, using more of the `IsEnabled` code paths. No user-facing implications. :heavy_check_mark: 
- The second (8d83f29) is just the result of calling [modernize](golang.org/x/tools/gopls/internal/analysis/modernize/cmd/modernize), so it should have no implications :heavy_check_mark: 
- The third (aa9397d) just removes dead code and replaces custom rounding in tests with `math.Round`. So hopefully no implications either :heavy_check_mark: 
- The fourth (91a442f) is a bit more interesting: it switches to `defer` for some unlocking since it has been pretty on-par performance-wise [for a long time](https://go-review.googlesource.com/c/go/+/202340), and is marginally safer. It also replaces `math/rand` with `math/rand/v2`, just out of force of habit :see_no_evil: It _should_ have no sensible implications for consumers, but may require a bit more scrutiny. :detective: 
- The fifth (9f05cd0) drops `reflect` usage in the internal `every` function. It causes a measurable performance improvement (see benchmarks in OP) and should have no side-effects, since it implements basically the same checks as before, just on compile time.
- The sixth (9becd32) could also use some scrutiny: we now bubble any errors during storage reset. I believe this is safer than ignoring the error, but I may be overlooking some side-effects. :detective: 
- And lastly the seventh (4912388) is a pretty straightforward micro-optimization and just pre-allocates the `allConstraints` slice to the worst-case larges capacity. Seems pretty safe? :heavy_check_mark: